to_do_app/requirements.txt
api==0.0.7
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.13
django-filter==24.3
djangorestframework==3.15.2
idna==3.7
importlib_metadata==8.2.0
Markdown==3.6
nose==1.3.7
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-json-logger==2.0.7
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
users==1.0.dev0
zipp==3.19.2
